+++ billeteravirtual/src/main/java/ar/com/ada/api/billeteravirtual/models/request/MovimientoRequest.java
@@ -1,5 +1,7 @@
 package ar.com.ada.api.billeteravirtual.models.request;
 
+import java.math.BigDecimal;
+
 /**
  * MovimientoRequest
  */
@@ -7,7 +9,7 @@ public class MovimientoRequest {
 
     //public String emailDestinatario;
     public String concepto;
-    public double importe;
+    public BigDecimal importe;
     public String emailDestinatario;
     public String moneda;
